video-proxy/requirements.txt
Flask==3.0.0
google-cloud-storage==2.10.0
gunicorn==21.2.0
werkzeug==3.0.0